refactor(web): 복잡한 상수 관심사 분리
