--- apps/web/src/components/account/accountCategory/index.ts
@@ -1,1 +1,0 @@
-export { default } from './accountCategory';
