+++ apps/web/src/components/account/accountCategoryList/index.ts
@@ -1,0 +1,1 @@
+export { default } from './accountCategoryList';
