+++ apps/web/src/components/account/index.ts
@@ -1,4 +1,4 @@
-export { default as AccountCategory } from './accountCategory';
+export { default as AccountCategoryList } from './accountCategoryList';
 export { default as AccountUpdateForm } from './accountUpdateForm';
 export { default as AccountPasswordForm } from './accountPasswordForm';
 export { default as Profile } from './profile';
