+++ apps/web/src/components/inquiry/partnershipForm/partnershipForm.tsx
@@ -1,7 +1,6 @@
 import { ErrorCode, fetcher } from '@supercarmarket/lib';
 import { Alert, Button, Form, Wrapper } from '@supercarmarket/ui';
 import { Modal } from 'components/common/modal';
-import inquiry, { InquiryPartnershipFormState } from 'constants/inquiry';
 import ModalContext from 'feature/modalContext';
 import { useSession } from 'next-auth/react';
 import * as React from 'react';
@@ -11,6 +10,7 @@ import { useRouter } from 'next/navigation';
 
 import InquiryFormItem from '../inquiryFormItem';
 import { useDebounce } from '@supercarmarket/hooks';
+import { form, type FormState } from 'constants/form/partnership';
 
 const formatter = (category: string) => {
   if (category === '전체') return 'ALL';
@@ -26,7 +26,7 @@ const PartnershipForm = () => {
   const [error, setError] = React.useState<string | null>(null);
   const { onClose, onOpen } = React.useContext(ModalContext);
   const { replace } = useRouter();
-  const methods = useForm<InquiryPartnershipFormState>();
+  const methods = useForm<FormState>();
 
   const handleModal = React.useCallback(
     (href: string) => {
@@ -37,9 +37,9 @@ const PartnershipForm = () => {
     [onClose, replace]
   );
 
-  const handleRequire = async (data: InquiryPartnershipFormState) => {
+  const handleRequire = async (data: FormState) => {
     Object.entries(data).map(([key, value]) => {
-      const target = key as keyof InquiryPartnershipFormState;
+      const target = key as keyof FormState;
 
       if (target == 'address' && !value.length) {
         methods.setError(target, {
@@ -76,7 +76,7 @@ const PartnershipForm = () => {
   };
 
   const debouncedSubmit = useDebounce(
-    async (data: InquiryPartnershipFormState) =>
+    async (data: FormState) =>
       handleRequire(data).then(async () => {
         setError(null);
 
@@ -152,12 +152,8 @@ const PartnershipForm = () => {
           gap: 24px;
         `}
       >
-        {inquiry.register.partnership.map((d) => (
-          <InquiryFormItem
-            key={d.htmlFor}
-            callback={(d) => console.log(d)}
-            {...d}
-          />
+        {form.map((d) => (
+          <InquiryFormItem key={d.htmlFor} {...d} />
         ))}
         <Wrapper.Item
           css={css`
