+++ apps/web/src/constants/link/index.ts
@@ -1,0 +1,9 @@
+import { type UrlObject } from 'url';
+
+export type Href = string | UrlObject;
+
+export type Links = {
+  title: string;
+  href: Href;
+  children?: Links[];
+};
