+++ apps/web/src/pages/market/index.tsx
@@ -16,7 +16,7 @@ import layout from 'components/layout';
 import MarketBanner from 'components/market/marketBanner';
 import MarketCar from 'components/market/marketCar';
 import MarketFilter from 'components/market/marketFilter';
-import { CATEGORY, CATEGORY_VALUES, MARKET_LINKS } from 'constants/market';
+import { CATEGORY, CATEGORY_VALUES } from 'constants/market';
 import type { InferGetServerSidePropsType, NextPageContext } from 'next/types';
 import * as React from 'react';
 import { ErrorBoundary } from 'react-error-boundary';
@@ -25,6 +25,7 @@ import { css } from 'styled-components';
 import HeadSeo from 'components/common/headSeo';
 import Advertisement from 'components/common/advertisement';
 import { prefetchMarket, QUERY_KEYS } from 'http/server/market';
+import { linsk } from 'constants/link/market';
 
 const MarketFilterPage: NextPageWithLayout = ({
   category,
@@ -93,7 +94,7 @@ const MarketFilterPage: NextPageWithLayout = ({
                 onReset={reset}
                 fallbackRender={(props) => <ErrorFallback {...props} />}
               >
-                <Category links={MARKET_LINKS} category={category} />
+                <Category links={linsk} category={category} />
                 <MarketFilter />
                 <MarketCar />
               </ErrorBoundary>
